+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Install dependencies via uv from the lockfile in the Dockerfile and the pylint CI workflow
 - Drop support for Python older than 3.12; require Python >= 3.12
 - Add support for the AWS EU Sovereign Cloud (aws-eusc) partition (upstream #239)
+- Add support for the AWS China (aws-cn) partition (upstream #182)
 
 ## [0.5.5] 2024-03-19
 - Bugfix [#199](https://github.com/okta-awscli/okta-awscli/issues/199) duplicates of data inside config file